+++ fuzz.py
@@ -10,7 +10,7 @@ import re
 import hypothesmith
 from hypothesis import HealthCheck, given, settings, strategies as st
 
-import black
+import blackish
 from blib2to3.pgen2.tokenize import TokenError
 
 
@@ -29,7 +29,7 @@ from blib2to3.pgen2.tokenize import TokenError
     src_contents=hypothesmith.from_grammar() | hypothesmith.from_node(),
     # Using randomly-varied modes helps us to exercise less common code paths.
     mode=st.builds(
-        black.FileMode,
+        blackish.FileMode,
         line_length=st.just(88) | st.integers(0, 200),
         string_normalization=st.booleans(),
         preview=st.booleans(),
@@ -38,16 +38,16 @@ from blib2to3.pgen2.tokenize import TokenError
     ),
 )
 def test_idempotent_any_syntatically_valid_python(
-    src_contents: str, mode: black.FileMode
+    src_contents: str, mode: blackish.FileMode
 ) -> None:
     # Before starting, let's confirm that the input string is valid Python:
     compile(src_contents, "<string>", "exec")  # else the bug is in hypothesmith
 
     # Then format the code...
     try:
-        dst_contents = black.format_str(src_contents, mode=mode)
-    except black.InvalidInput:
-        # This is a bug - if it's valid Python code, as above, Black should be
+        dst_contents = blackish.format_str(src_contents, mode=mode)
+    except blackish.InvalidInput:
+        # This is a bug - if it's valid Python code, as above, Blackish should be
         # able to cope with it.  See issues #970, #1012
         # TODO: remove this try-except block when issues are resolved.
         return
@@ -56,17 +56,17 @@ def test_idempotent_any_syntatically_valid_python(
             e.args[0] == "EOF in multi-line statement"
             and re.search(r"\\($|\r?\n)", src_contents) is not None
         ):
-            # This is a bug - if it's valid Python code, as above, Black should be
+            # This is a bug - if it's valid Python code, as above, Blackish should be
             # able to cope with it.  See issue #1012.
             # TODO: remove this block when the issue is resolved.
             return
         raise
 
     # And check that we got equivalent and stable output.
-    black.assert_equivalent(src_contents, dst_contents)
-    black.assert_stable(src_contents, dst_contents, mode=mode)
+    blackish.assert_equivalent(src_contents, dst_contents)
+    blackish.assert_stable(src_contents, dst_contents, mode=mode)
 
-    # Future test: check that pure-python and mypyc versions of black
+    # Future test: check that pure-python and mypyc versions of blackish
     # give identical output for identical input?